[add: golem] 신규 몬스터 추가 및 파일 정리

+++ TP02/TP02/Enemy/Enemy.h
@@ -18,5 +18,5 @@ public:
 
 protected:
 	bool m_bIsDead;
-	int m_iExpReward;
+	float m_iExpReward;
 };
